+++ CHANGELOG.md
@@ -56,6 +56,8 @@
 
 ## stream-chat-android-ui-components
 ### 🐞 Fixed
+- Fixed the crash happening while editing a message with a recording attachment. [#5220](https://github.com/GetStream/stream-chat-android/pull/5220)
+- Fixed intermittent crash when opening a channel. [#5219](https://github.com/GetStream/stream-chat-android/pull/5219)
 
 ### ⬆️ Improved
 
@@ -67,10 +69,15 @@
 
 ## stream-chat-android-compose
 ### 🐞 Fixed
+- Fixed `MessageListViewModel.selectExtendedReactions` not calling the proper function in `MessageListController`. [#5218](https://github.com/GetStream/stream-chat-android/pull/5218)
 
 ### ⬆️ Improved
 
 ### ✅ Added
+- Added `MessageTextFormatter` to format the message text. [#5214](https://github.com/GetStream/stream-chat-android/pull/5214)
+  * Can be overridden by `ChatTheme.messageTextFormatter`
+- Added `QuotedMessageTextFormatter` to format the quoted message text. [#5214](https://github.com/GetStream/stream-chat-android/pull/5214)
+  * Can be overridden by `ChatTheme.quotedMessageTextFormatter` 
 
 ### ⚠️ Changed
 
@@ -86,6 +93,15 @@
 ### ⚠️ Changed
 
 ### ❌ Removed
+
+# March 20th, 2024 - 6.2.1
+## stream-chat-android-ui-components
+### ⬆️ Improved
+- Added `flag-message` capability support. [#5211](https://github.com/GetStream/stream-chat-android/pull/5211)
+
+## stream-chat-android-compose
+### ✅ Added
+- Added `flag-message` capability support. [#5211](https://github.com/GetStream/stream-chat-android/pull/5211)
 
 # March 15th, 2024 - 6.2.0
 ## Common changes for all artifacts